Move TrieMetadata to Libplanet.Store
